feat: Rename shared utils
