+++ src/utils/index.ts
@@ -1,2 +1,2 @@
 export { default as Dummy } from './dummy';
-export { isEmail, isPassword, isSame } from './validation';
+export { isEmail, isPassword, isSame } from './shared';
